+++ connectors/src/connectors/microsoft_bot/bot_messaging_utils.ts
